Button is now stateless (uses mouse tracker HOC)
